Rename table ip2nationCountries
Honestly, I don't know what this table is doing in world, because I think it is in auth, or should be in it. Anyway, for the moment, we are just going to rename it, so that it works on both Windows and Linux. Since Windows is not case sensitive, but Linux is case sensitive.